requirements.txt
Flask
opencv-python
numpy
werkzeug
gunicorn
tensorflow-intel-0.0.1 